requirements.txt
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
arrow==1.3.0
asgiref==3.8.1
asteval==0.9.33
astroid==3.3.8
attrs==23.2.0
black==23.11.0
build==1.0.3
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.11.17
CFGraph==0.2.1
chardet==5.2.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
colorama==0.4.6
crashtest==0.4.1
curies==0.7.10
deepdiff==6.7.1
Deprecated==1.2.14
dill==0.3.9
distlib==0.3.7
Django==5.1.5
django-cors-headers==4.6.0
django-sslserver==0.22
djangorestframework==3.15.2
dulwich==0.21.6
environs==14.1.0
et_xmlfile==2.0.0
exceptiongroup==1.1.3
fastjsonschema==2.19.0
filelock==3.13.1
flexcache==0.3
flexparser==0.3.1
fqdn==1.5.1
google-auth==2.23.4
graphviz==0.20.3
greenlet==3.1.1
gunicorn==23.0.0
hbreader==0.9.1
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
installer==0.7.0
isodate==0.6.1
isoduration==20.11.0
isort==5.13.2
jaraco.classes==3.3.0
Jinja2==3.1.4
json-flattener==0.1.9
jsonasobj==1.3.1
jsonasobj2==1.0.4
jsonpatch==1.33
jsonpath-ng==1.7.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keyring==24.3.0
keyrings.google-artifactregistry-auth==1.1.1
lark==1.1.9
linkml==1.8.5
linkml-dataops==0.1.0
linkml-map==0.3.7
linkml-runtime==1.8.3
MarkupSafe==3.0.2
marshmallow==3.26.0
mccabe==0.7.0
more-itertools==10.1.0
msgpack==1.0.7
mypy-extensions==1.0.0
openpyxl==3.1.5
ordered-set==4.1.0
packaging==23.2
parse==1.20.2
pathspec==0.12.1
pexpect==4.8.0
Pint==0.24.1
pkginfo==1.9.6
platformdirs==3.11.0
pluggy==1.3.0
ply==3.11
poetry==1.7.0
poetry-core==1.8.1
poetry-plugin-export==1.6.0
prefixcommons==0.1.12
prefixmaps==0.2.4
psycopg2==2.9.10
ptyprocess==0.7.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.8.2
pydantic_core==2.20.1
PyJSG==0.11.10
pylint==3.3.3
pyparsing==3.1.2
pyproject_hooks==1.0.0
PyShEx==0.8.1
PyShExC==0.9.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-logging==2015.11.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyTrie==0.4.0
pywin32-ctypes==0.2.2
PyYAML==6.0.1
rapidfuzz==3.5.2
rdflib==7.0.0
rdflib-jsonld==0.6.1
rdflib-shim==1.0.3
referencing==0.35.1
requests==2.31.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3987==1.3.8
rpds-py==0.19.0
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
shellingham==1.5.4
ShExJSG==0.8.2
six==1.16.0
sortedcontainers==2.4.0
sparqlslurper==0.5.1
SPARQLWrapper==2.0.0
SQLAlchemy==2.0.36
sqlparse==0.5.3
tomli==2.0.1
tomlkit==0.12.3
trove-classifiers==2023.11.14
types-python-dateutil==2.9.0.20241003
typing_extensions==4.9.0
tzdata==2025.1
ucumvert==0.1.2
uri-template==1.3.0
urllib3==2.1.0
virtualenv==20.24.7
watchdog==6.0.0
webcolors==24.8.0
wrapt==1.16.0
zipp==3.17.0
